Organize cloud sign-in settings files
